Updated Build
The previous build ran at an uncapped frame rate which caused many game breaking bugs such as AI clipping through walls and incredibly low sensitivity. This new build has an fps lock which can't be found in the unity files right now as the build was made a long time ago when I got a new and improved pc which ran the game at almost 500 fps, I lost those files but all it really was is a single script with 2 lines of code which turned vsync off and set a frame rate cap.